requirements.txt
discord.py
google-genai
dotenv
pymongo